♻ Refactor: 제미나이 코드리뷰 반영
